backend/requirements.txt
asyncpg==0.30.0
databases[asyncpg]==0.9.0
faiss-cpu==1.8.0.post1
fastapi[standard]==0.113.0
langchain-community==0.2.19
langchain-ollama==0.1.3
nomic[local]==3.4.1
ollama==0.4.7
psycopg2-binary==2.9.10
pymilvus==2.5.6
pymilvus[model]
pydantic==2.9.0
uvicorn==0.27.1
python-dotenv==1.1.0
